COMPASS v5.4.4 hotfix : - Debug KL2V when DM has unseen actuators

+++ shesha/util/make_pupil.py
@@ -1,7 +1,7 @@
 ## @package   shesha.util.make_pupil
 ## @brief     Pupil creation functions
 ## @author    COMPASS Team <https://github.com/ANR-COMPASS>
-## @version   5.4.3
+## @version   5.4.4
 ## @date      2022/01/24
 ## @copyright GNU Lesser General Public License
 #
@@ -692,8 +692,8 @@ def fillPolygon(x, y, i0, j0, scale, gap, N, index=0):
     # Last triangle has a special treatment because it crosses the axis
     # with theta=0=2pi
     n = x.shape[0]  # number of corners of polygon
-    indx, indy = (np.array([], dtype=np.int32), np.array([], dtype=np.int32))
-    distedge = np.array([], dtype=np.float32)
+    indx, indy = (np.array([], dtype=np.int), np.array([], dtype=np.int))
+    distedge = np.array([], dtype=np.float)
     for i in range(n):
         j = i + 1  # j=element next i except when i==n : then j=0 (cycling)
         if j == n:
